multi: remove duplication by moving HTLC to rfqmsg
We had to introduce some code duplication in order to avoid circular
package dependencies. We can resolve both by moving the HTLC and
AssetBalance structs from the tapchannelmsg to the rfqmsg package.
