+++ tapchannelmsg/records.go
@@ -512,125 +512,6 @@ func DecodeCommitment(blob tlv.Blob) (*Commitment, error) {
 	return &c, nil
 }
 
-// NoneRfqIDRecord creates an optional record that represents a None RFQ ID.
-func NoneRfqIDRecord() tlv.OptionalRecordT[HtlcRfqIDType, rfqmsg.ID] {
-	return tlv.OptionalRecordT[HtlcRfqIDType, rfqmsg.ID]{
-		Option: lfn.None[tlv.RecordT[HtlcRfqIDType, rfqmsg.ID]](),
-	}
-}
-
-// SomeRfqIDRecord creates an optional record that represents an RFQ ID.
-func SomeRfqIDRecord(
-	id rfqmsg.ID) tlv.OptionalRecordT[HtlcRfqIDType, rfqmsg.ID] {
-
-	return tlv.SomeRecordT(
-		tlv.NewPrimitiveRecord[HtlcRfqIDType, rfqmsg.ID](id),
-	)
-}
-
-// Htlc is a record that represents the capacity change related to an in-flight
-// HTLC. This entails all the (asset_id, amount) tuples and other information
-// that we may need to be able to update the TAP portion of a commitment
-// balance.
-type Htlc struct {
-	// Amounts is a list of asset balances that are changed by the HTLC.
-	Amounts tlv.RecordT[HtlcAmountRecordType, AssetBalanceListRecord]
-
-	// RfqID is the RFQ ID that corresponds to the HTLC.
-	RfqID tlv.OptionalRecordT[HtlcRfqIDType, rfqmsg.ID]
-}
-
-// NewHtlc creates a new Htlc record with the given funded assets.
-func NewHtlc(amounts []*AssetBalance, rfqID fn.Option[rfqmsg.ID]) *Htlc {
-	htlc := &Htlc{
-		Amounts: tlv.NewRecordT[HtlcAmountRecordType](
-			AssetBalanceListRecord{
-				Balances: amounts,
-			},
-		),
-	}
-	rfqID.WhenSome(func(id rfqmsg.ID) {
-		htlc.RfqID = SomeRfqIDRecord(id)
-	})
-
-	return htlc
-}
-
-// Balances returns the list of asset Balances that are updated in the Htlc
-// struct.
-func (h *Htlc) Balances() []*AssetBalance {
-	return h.Amounts.Val.Balances
-}
-
-// Records returns the records that make up the Htlc.
-func (h *Htlc) Records() []tlv.Record {
-	records := []tlv.Record{
-		h.Amounts.Record(),
-	}
-
-	h.RfqID.WhenSome(func(r tlv.RecordT[HtlcRfqIDType, rfqmsg.ID]) {
-		records = append(records, r.Record())
-	})
-
-	return records
-}
-
-// Encode serializes the Htlc to the given io.Writer.
-func (h *Htlc) Encode(w io.Writer) error {
-	tlvRecords := h.Records()
-
-	// Create the tlv stream.
-	tlvStream, err := tlv.NewStream(tlvRecords...)
-	if err != nil {
-		return err
-	}
-
-	return tlvStream.Encode(w)
-}
-
-// Decode deserializes the Htlc from the given io.Reader.
-func (h *Htlc) Decode(r io.Reader) error {
-	rfqID := h.RfqID.Zero()
-
-	// Create the tlv stream.
-	tlvStream, err := tlv.NewStream(
-		h.Amounts.Record(),
-		rfqID.Record(),
-	)
-	if err != nil {
-		return err
-	}
-
-	typeMap, err := tlvStream.DecodeWithParsedTypesP2P(r)
-	if err != nil {
-		return err
-	}
-
-	if val, ok := typeMap[h.RfqID.TlvType()]; ok && val == nil {
-		h.RfqID = tlv.SomeRecordT(rfqID)
-	}
-
-	return nil
-}
-
-// Bytes returns the serialized Htlc record.
-func (h *Htlc) Bytes() []byte {
-	var buf bytes.Buffer
-	_ = h.Encode(&buf)
-	return buf.Bytes()
-}
-
-// DecodeHtlc deserializes a Htlc from the given blob.
-func DecodeHtlc(blob tlv.Blob) (*Htlc, error) {
-	var h Htlc
-	err := h.Decode(bytes.NewReader(blob))
-	if err != nil {
-		return nil, err
-	}
-
-	return &h, nil
-}
-
 // HtlcAuxLeaf is a record that represents the auxiliary leaf of an HTLC and
 // the optional second level leaf. The second level leaf is optional because it
 // is not set in every case of the HTLC creation flow.
@@ -890,88 +771,10 @@ func dHtlcAuxLeafMapRecord(r io.Reader, val interface{}, buf *[8]byte,
 	)
 }
 
-// AssetBalance is a record that represents the amount of an asset that is
-// being transferred or is available to be spent.
-type AssetBalance struct {
-	// AssetID is the ID of the asset that this output is associated with.
-	AssetID tlv.RecordT[tlv.TlvType0, asset.ID]
-
-	// Amount is the amount of the asset that this output represents.
-	Amount tlv.RecordT[tlv.TlvType1, uint64]
-}
-
-// NewAssetBalance creates a new AssetBalance record with the given asset ID and
-// amount.
-func NewAssetBalance(assetID asset.ID, amount uint64) *AssetBalance {
-	return &AssetBalance{
-		AssetID: tlv.NewRecordT[tlv.TlvType0](assetID),
-		Amount:  tlv.NewPrimitiveRecord[tlv.TlvType1](amount),
-	}
-}
-
-// records returns the records that make up the AssetBalance.
-func (a *AssetBalance) records() []tlv.Record {
-	return []tlv.Record{
-		a.AssetID.Record(),
-		a.Amount.Record(),
-	}
-}
-
-// Encode serializes the AssetBalance to the given io.Writer.
-func (a *AssetBalance) Encode(w io.Writer) error {
-	tlvRecords := a.records()
-
-	// Create the tlv stream.
-	tlvStream, err := tlv.NewStream(tlvRecords...)
-	if err != nil {
-		return err
-	}
-
-	return tlvStream.Encode(w)
-}
-
-// Decode deserializes the AssetBalance from the given io.Reader.
-func (a *AssetBalance) Decode(r io.Reader) error {
-	// Create the tlv stream.
-	tlvStream, err := tlv.NewStream(a.records()...)
-	if err != nil {
-		return err
-	}
-
-	return tlvStream.Decode(r)
-}
-
-// Sum returns the sum of the amounts of all the asset Balances in the list.
-func Sum(balances []*AssetBalance) uint64 {
-	var sum uint64
-	for _, balance := range balances {
-		sum += balance.Amount.Val
-	}
-	return sum
-}
-
-// Bytes returns the serialized AssetBalance record.
-func (a *AssetBalance) Bytes() []byte {
-	var buf bytes.Buffer
-	_ = a.Encode(&buf)
-	return buf.Bytes()
-}
-
-// DecodeAssetBalance deserializes a AssetBalance from the given blob.
-func DecodeAssetBalance(blob tlv.Blob) (*AssetBalance, error) {
-	var a AssetBalance
-	err := a.Decode(bytes.NewReader(blob))
-	if err != nil {
-		return nil, err
-	}
-
-	return &a, nil
-}
-
 // AssetOutput is a record that represents a single asset UTXO.
 type AssetOutput struct {
 	// AssetBalance is the asset ID and amount of the output.
-	AssetBalance
+	rfqmsg.AssetBalance
 
 	// Proof is the last transition proof that proves this output was
 	// committed to in the Bitcoin transaction that anchors this asset
@@ -985,7 +788,7 @@ func NewAssetOutput(assetID asset.ID, amount uint64,
 	p proof.Proof) *AssetOutput {
 
 	return &AssetOutput{
-		AssetBalance: AssetBalance{
+		AssetBalance: rfqmsg.AssetBalance{
 			AssetID: tlv.NewRecordT[tlv.TlvType0](assetID),
 			Amount:  tlv.NewPrimitiveRecord[tlv.TlvType1](amount),
 		},
@@ -1209,130 +1012,6 @@ func dHtlcAssetOutput(r io.Reader, val interface{}, buf *[8]byte,
 	return tlv.NewTypeForEncodingErr(
 		val, "map[input.HtlcIndex]AssetOutputListRecord",
 	)
-}
-
-// AssetBalanceListRecord is a record that represents a list of asset Balances.
-type AssetBalanceListRecord struct {
-	Balances []*AssetBalance
-}
-
-// Sum returns the sum of the amounts of all the asset Balances in the list.
-func (l *AssetBalanceListRecord) Sum() uint64 {
-	return Sum(l.Balances)
-}
-
-// Record creates a Record out of a AssetBalanceListRecord using the
-// eAssetBalanceList and dAssetBalanceList functions.
-//
-// NOTE: This is part of the tlv.RecordProducer interface.
-func (l *AssetBalanceListRecord) Record() tlv.Record {
-	size := func() uint64 {
-		var (
-			buf     bytes.Buffer
-			scratch [8]byte
-		)
-		err := eAssetBalanceList(&buf, &l.Balances, &scratch)
-		if err != nil {
-			panic(err)
-		}
-
-		return uint64(buf.Len())
-	}
-
-	// Note that we set the type here as zero, as when used with a
-	// tlv.RecordT, the type param will be used as the type.
-	return tlv.MakeDynamicRecord(
-		0, &l.Balances, size, eAssetBalanceList, dAssetBalanceList,
-	)
-}
-
-// Encode serializes the AssetBalanceListRecord to the given io.Writer.
-func (l *AssetBalanceListRecord) Encode(w io.Writer) error {
-	// Create the tlv stream.
-	tlvStream, err := tlv.NewStream(l.Record())
-	if err != nil {
-		return err
-	}
-
-	return tlvStream.Encode(w)
-}
-
-// Decode deserializes the AssetBalanceListRecord from the given io.Reader.
-func (l *AssetBalanceListRecord) Decode(r io.Reader) error {
-	// Create the tlv stream.
-	tlvStream, err := tlv.NewStream(l.Record())
-	if err != nil {
-		return err
-	}
-
-	return tlvStream.Decode(r)
-}
-
-// eAssetBalanceList is an encoder for AssetBalanceListRecord.
-func eAssetBalanceList(w io.Writer, val interface{}, buf *[8]byte) error {
-	if v, ok := val.(*[]*AssetBalance); ok {
-		numBalances := uint64(len(*v))
-		if err := tlv.WriteVarInt(w, numBalances, buf); err != nil {
-			return err
-		}
-		var outputBuf bytes.Buffer
-		for _, balance := range *v {
-			if err := balance.Encode(&outputBuf); err != nil {
-				return err
-			}
-			balanceBytes := outputBuf.Bytes()
-			err := asset.InlineVarBytesEncoder(
-				w, &balanceBytes, buf,
-			)
-			if err != nil {
-				return err
-			}
-			outputBuf.Reset()
-		}
-		return nil
-	}
-	return tlv.NewTypeForEncodingErr(val, "[]*AssetBalance")
-}
-
-// dAssetBalanceList is a decoder for AssetBalanceListRecord.
-func dAssetBalanceList(r io.Reader, val interface{}, buf *[8]byte,
-	_ uint64) error {
-
-	if typ, ok := val.(*[]*AssetBalance); ok {
-		numBalances, err := tlv.ReadVarInt(r, buf)
-		if err != nil {
-			return err
-		}
-
-		// Avoid OOM by limiting the number of balances we accept.
-		if numBalances > MaxNumOutputs {
-			return fmt.Errorf("%w: too many balances",
-				ErrListInvalid)
-		}
-
-		if numBalances == 0 {
-			return nil
-		}
-
-		outputs := make([]*AssetBalance, numBalances)
-		for i := uint64(0); i < numBalances; i++ {
-			var outputBytes []byte
-			err := asset.InlineVarBytesDecoder(
-				r, &outputBytes, buf, tlv.MaxRecordSize,
-			)
-			if err != nil {
-				return err
-			}
-			outputs[i] = &AssetBalance{}
-			err = outputs[i].Decode(bytes.NewReader(outputBytes))
-			if err != nil {
-				return err
-			}
-		}
-		*typ = outputs
-		return nil
-	}
-	return tlv.NewTypeForEncodingErr(val, "[]*AssetBalance")
 }
 
 // AssetOutputListRecord is a record that represents a list of asset outputs.
